+++ src/Migration.jl
@@ -570,5 +570,3 @@ const drop_sequence = remove_sequence
 const dropsequence = remove_sequence
 
 end
-
-const Migrations = Migration