+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## v0.12.1 - 2019-08-29
 
 * SearchLight environment inherits Genie environment, if available
+* Improved handling of 2006 and 2013 MySQL errors (connection lost)
 * fixes an issue where reading the configuration file left the file open
-* API cleanup
+* API cleanup in `SearchLight.Database`
 
 ## v0.12.0 - 2019-08-29